This is a logistics model coding file
This file is for amazon. And it is aimed to find to rank of the factor that is important for AMZ to attract sellers to create a business on AMZ. The data is collected from desk. Factors that need to be chosen are Current cycle sales,  Number of associated videos,  Unit price,  Related influencers,  Related live streaming,     Number of historical products,  Monthly estimated revenue (million).